// src/store/types/HI/TQuickQuote.ts
export type TQuickQuote = {
  premiumDetails: TPremiumDetails;
  productDetails: {
    id: string;
    parent_id: string;
    product_type_id: string;
    product_type_code: string;
    product_code: string;
    product_name: string;
    product_desc: string;
    valid_from: string;
    valid_to: string;
    rating: string;
    brochure: string;
    quote_action: string;
    rate_calc_method: string;
    proposal_method: string;
    is_online: string;
    status: string;
    created: string;
    tax_calc_method: string;
    discount_calc_method: string;
    buy_online_code: string;
    special_feature_codes: string;
    policy_covers: string;
    bs_ty_new: string;
    bs_ty_rollover: string;
    bs_ty_exp: string;
    short_name: string;
    prem_calc_action: string;
    tax_calc_action: string;
    discount_calc_action: string;
  };
  CompanyDetails: {
    id: string;
    company_master_id: string;
    company_code: string;
    valid_from: string;
    valid_to: string;
    product_type_id: string;
    product_type_code: string;
    claim_ratio: string;
    avg_time_to_sttle_claim: string;
    net_income_ration: string;
    policy_sales_growth: string;
    customer_satisfaction: string;
    logo: string;
    status: string;
    created: string;
    logoUrl: string;
  };
  featureCount: number;
  exclusionCount: number;
};

export type TPremiumDetails = {
  sumInsured: string;
  buildingSI: string;
  householdSI: string;
  enquiryId: string;
  netPremium: string;
  serviceTax: string;
  grossPremium: string;
  discountPremium: string;
  discountedNetPremium: string;
  earthQuakePremium: string;
  terrorismPremium: string;
  lossOfRentPremium: string;
  additionalPremium: string;
  terrorismDiscountPremium: string;
  earthquakeDiscountPremium: string;
  otherDiscountPremium: string;
  ncbDiscount: string;
  odDiscount: string;
  antiTheftDiscount: string;
  associationDiscount: string;
  ePolicyDiscount: string;
  otherDiscount: string;
  deductibleDiscount: string;
  ambulanceChargesCover: string;
  consumablesCover: string;
  passengerCover: string;
  digit_plan_type: string;
  planCode: string | null;
  derivedPremium: string | null;
  totalPremium: string | null;
  totalTax: string | null;
  basicPremium: string | null;
  correlationId?: string;
  // grossPremium?: string | null;
};